cloud_storage_gateway/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
huaweicloudsdkcore
huaweicloudsdkiam
esdk-obs-python
python-dotenv
requests
